requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
python-dotenv
jinja2==3.1.2
aiofiles==23.1.0
python-multipart
groq>=0.33.0
itsdangerous==2.1.2
starlette
PyPDF2>=3.0.0
requests
